requirements.txt
asgiref==3.5.2
autopep8==1.5.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.44.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-js-asset==2.0.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-storages==1.12.3
django-summernote==0.8.11.6
django-tinymce==3.5.0
djangorestframework==3.12.0
drf-yasg==1.20.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.0.1
psycopg2-binary==2.9.4
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.14
